chore(CircleCI): remove yarn run v1.22.10 $ react-scripts build Creating an optimized production build... Compiled successfully.
File sizes after gzip:

  132.85 KB  build/static/js/2.8c6a256d.chunk.js
  23.25 KB   build/static/css/2.4c97ca4f.chunk.css
  8.31 KB    build/static/js/main.ee24dc21.chunk.js
  1.56 KB    build/static/js/3.f78ba714.chunk.js
  1.16 KB    build/static/js/runtime-main.06c3411c.js

The project was built assuming it is hosted at /.
You can control this with the homepage field in your package.json.

The build folder is ready to be deployed.
You may serve it with a static server:

  yarn global add serve
  serve -s build

Find out more about deployment here:

  https://cra.link/deployment

Done in 14.55s. command from jobs