Resolved a merge conflict (App.tsx)
Rebase.

This is a combination of the following commits:

feat(gitignore and index): Moved the sentry dsn string into an environment file and added the file to gitignore

fix(gitignore): Fixed the path to the dsn file

feat(ErrorLogger): Created the ErrorLogger interface

feat(various): Refactored the ErrorLogger and added the basic versions of the Console/Sentry Loggers

feat(various): Now using the ErrorBoundary that pertains to the current environment
Production uses Sentry.ErrorBoundary whereas anything else uses the ConsoleErrorBoundary. Removed the SentryLogger since the sentry error boundary will log the error to sentry.

feat(various): Made the console logger log the error and errorInfo that is sent to its logError method
Also, moved the sentryDSN to the environment files and created a method to get the right dsn based on the current environment.

test(ConsoleErrorBoundary and etc): Created a test that confirms that the ConsoleErrorBoundary correctly logs errors that are thrown
Also, created the error-boundary feature folder and moved the error boundaries to it.

test(various): Added a basic fallback ui to the ConsoleErrorBoundary and confirmed that it is displayed via the test.
Also, simplified the logging of the error.

refactor(various): Removed unneeded code; Edited the eslint-disable lines; Made the AppErrorBoundary into a FC since it wasn't necessary for it to be a class component