Resolved a merge conflict.
Rebase. This combines the following commits:

Spike out notifications.

Reorganize notification stuff to be in their own feature.

SmallContainer

fix(various): Resolved some things that needed to be fixed after doing
the rebase

feat(BitwiseNavbar and useNavLink): Re-added a basic versin of the
notification bell and the count to the navbar

feat(useLiveNotifications): Added the reducer and its associated
interfaces and initial state const

fix(useLiveNotifications): Resolved the undefined side of using a
reducer and fixed one of the reducer's cases

feat(useLiveNotifications): Finished replacing the useState
implementation with the reducer approach

fix(BitwiseNavbar): Corrected the count for the notification button

Resolved a merge conflict.

Rebase. This combines the following commits:

feat(BitwiseNavbar): Started to update the notification indicator so
that it is consistent with the mockups (placing the number in a bubble
to the top right of the bell)

refactor(BitwiseNavbar): Refactored the previous approach; Now using the actual notification count; Set the cap to 9 based on the limited space that is available in the nav bar

feat(BitwiseNavbar): The count bubble will only appear when the number of notifications is more than zero.

feat(BitwiseNavbar): The NotificationButtonContainer now supports both the mobile and desktop perspectives.

fix(BitwiseNavbar): Fixed a responsiveness issue with the notification counter at the intersection of the mobile/desktop perspectives; Made it so that there would always be a pointer for the NotificationButtonContainer

fix(various): No longer hard-coding the colors for the notification counter; Made the counter look better when looking at it from the mobile perspective

refactor(BitwiseNavbar): Edited the styling for the NotificationButton so that unnecessary styles were removed and px values were changed to rem

Also, changed its z-index so that it is more consistent with Bootstrap's z-index levels.

useState -> useReducer

Seperate out user change aspect of useLiveNotifications

Move count into reducer state.