Resolved a merge conflict.
Rebase. This includes the following commits:

feat(Layout and TopNav): Started to implement the TopNav component

feat(page and TopNav): The TopNav now looks like a top navigation bar.

feat(Layout and page): Made the navbar look a bit better and made sure it wasn't covering the Main content while in the desktop perspective

feat(various): Made the BitwiseNavBar support both the mobile and desktop persective; Moved the theme toggle to the left of the language selector for the desktop perspective

feat(various): Changed the background color of the navbar to a grey that's close to what's in the mockup for the desktop perspective

feat(various): Changed the light mode background for the NavUserDetails component for the desktop perspective

feat(Dropdown): Started to implement the dropdown that will show the Profile and Sign Out buttons

feat(various): Completed the first pass implementation for the Dropdown

feat(Dropdown): Made the Dropdown's background consistent with the background of the NavUserDetails component

feat(various): The container that surrounds the content of the listviews has now been expanded so that its content is aligned with the logo and the NavUserDetails component.

fix(Dropdown): Resolved a unique key warning

refactored(TopNav): Removed some unused code

refactor(various): Re-added the original version of the BitwiseNavbar with a slight change; I will be making it so that there is only one source of truth for the nav bar.

feat(NavBar and Layout): Now using the offcanvas-example navbar from the React-Boostrap documentation

feat(NavBar): Added the ThemeToggle, Language Selector, and the NavUserDetails components to the new Navbar

feat(various): Improved the placement of the theme, language, and nav user detail buttons in the offcanvas view

feat(NavBar): Added the previous nav links and made the mobile styling consistent with what we had originally

feat(NavBar): The sign out button will now only appear when the screen width is less than or equal to 767.

feat(various): The NavUserDetails dialog will now only show when the user is in the desktop perspective. Otherwise, clicking on the component will just direct the user to the UpdateUserProfile page.

feat(various): Made the navbar stay on top of all of the other elements; Changed the DataTable z-index stylings of the dropdown so that they are consistent with Bootstrap's conventions

fix(NavUserDetails): Fixed an issue where the dialog for the NavUserDetails component wasn't being shown in the right circumstances

feat(NavBar): Now setting the background color for the navbar based on the current theme and whether the user is in the desktop or mobile perspective

fix(NavBar): Fixed an issue with the background color in the offcanvas space when the dark theme is being used

feat(NavBar and NavUserDetails): Made the spacing around the NavUserDetails component look a bit better on mobile

fix(Dropdown): Fixed a theming issue that was occuring with the dark theme

refactor(various): Removed the Navbar and Nav components that were no longer needed; Updated references and removed console logs

refactor(utilities): Changed the z-index of the staging banner so that it would show on top of the navbar

refactor(Dropdown): Made a slight edit to the spacing between the right side of the opened dropdown and the right side of the screen

feat(BitwiseNavbar): The navbar will now have a marginTop of 56px when the app is in the staging environment.
By doing this, the staging banner won't show on top of it. Also, made a slight naming edit.

refactor(ToggleSwitch): Made slight edits to the media queries so that they would be consistent with the other ones that are associated with the navbar