Resolves a merge conflict
Rebase. This combines the following commits:

feat(UserProfilePicture and UpdateUserProfilePage): The UserProfilePicture component will now display a photo icon on top of the image when the showUploadPhotoButton prop is set to true.

refactor(UserProfilePicture): Created a styled component in order to remove the inline styling

feat(UserProfilePicture): Adds a tooltip that shows up when the user hovers over the camera icon

feat(various): The background and textColor of the tooltip will now be dependent on the current theme.

fix(UserProfilePicture): The UserProfilePicture component was being displayed incorrectly in the mobile perspective. This corrects that issue.

feat(various): Made some progress in testing how I would use a button to interact with the existing file input control code. When the photo is submitted, the associated formData is undefined. I will be looking into this on Monday.

fix(UpdateProfilePictureForm): Resolved the issue that was causing the
picture to not be submitted to the backend

feat(various): The user can now click on the camera icon, select their
image, and have their picture be uploaded and then shown to them.

Also, reverted the UserProfilePicture component.

feat(NewUpdateProfilePictureForm): Now showing a spinner when the user
is uploading a new profile picture. Also, made it so that the camera
icon's color depends on whether the user has a profilePicture or not.

refactor(various): Renamed the profilePictureBucket const; Removed the
unused consts from formState, Deleted the previous profile picture form
and corrected the file name of the new one

feat(various): feat(various): Made the spinner show up on top of the image and without the icon when a new image is being updated

refactor(various): Removed the inline styling and the container div from
LoadingSpinner; The place that uses the LoadingSpinner now has to define
its styled container.

refactor(NewLoadingSpinner): The NewLoadingSpinner component is no
longer necessary.

test(UpdateProfilePictureForm): Corrected the test and made necessary
edits to the associated form

refactor(various): Removed an unused prop and removed some console logs

refactor(UpdateProfilePictureForm): Renamed the updatingProfilePicture const

refactor(UpdateUserProfilePage): Deleted the row that was used to manage the user's profile photo and the code associated with it

fix(UpdateProfilePictureForm): Fixed the display of the error that would be displayed if the file input's accept prop were removed

test(UpdateProfilePictureForm.spec): Removed the test that was failing since we aren't testing toasts in other cases