feat(various): Now able to de-select a filter
Also, resolved a bug that was a part of the DataTableSearchAndFilters file. If
this wasn't edited, then it would also affect filters that had the same
operation. For example, if you had two filters that used the contains operation
and you removed one of them, it would remove both of them.