Rebase. This commits includes the following commits:
Resolved a merge conflict

Rebase. This includes the following commits:

feat(various): Re-added the ChangeLog component and the circular container
Did some slight edits to the circular container in order to remove the props that set the backgroundColor and dimensions.

feat(various): The ChangeLog is now online for the UpdateAgentView.

feat(various): Nearly done with adding the ChangeLog to the
UpdateUserView
I'm dealing with the situation where the historyUser of the admin user
is null. I'm assuming that is because the backend is what created this
user instead of this user.

fix(ChangeLog): Resolves the scenario where the historyUser property is
set to null

refactor(various): Removed the unnecessary console logs and commented out the console logs that are within the handleShowAllChanges methods in both views

refactor(various): Removed the duplicate isObject method; Moved the isStringOrNumber method; Refactored the location of the ChangeLog's methods

refactor(ChangeLog): Used moment in place of the getDateStr and getTimeStr methods

refactor(ChangeLog): Removed the need for the getActionInPastTense method

refactor(various): Created a HistoricalRecord generic type for the change items; Removed the defaultPropertyAccessor and defaultGetChangeItemAction methods (and related code) since they were no longer necessary

fix(ChangeLog): Corrected a typo

refactor(ChangeLog): Consolidated the getIconForAction and getChangeNameForAction methods into a single method

refactor(ChangeLog): Renamed some things in order to add clarity

refactor(ChangeLog): Slight edit

feat(UpdateUserView): Added the modal

feat(various): Completed the first pass for the useInfiniteLoading hook
Issue: The last page isn't loading for some reason.

feat(various): Changed the useInfiniteLoading hook so that it takes an
initial url instead of the request's parameters.
This also resolved that issue that was causing the last page not to
load.

refactor(various): Removed code that was no longer being used

refactor(various): Created a ChangeListGroup component so that the
rendering of the change list items can be used by both the modal and the
ChangeLog itself
Issue: The app crashes when you click on the Show All Changes button.
The styled editor name component is causing an error for some reason.

feat(UpdateUserView): The Load More button in the modal will only appear
when there are more items that can be loaded.

fix(ChangeListGroup): Found a fix for the issue that was causing a
nested styled component to not work in the appropriate manner

feat(various): Now limiting the amount of items that are displayed by
the ChangeLog

feat(ChangeLog): The Show All Changes button is now dependent on whether
the totalChanges is greater than the previewSize instead of the length
of the changeList.

feat(various): Made the Load More button's loading prop depend on the
isFetching const from the passed in query for the useInfiniteLoading
hook. Also, centered the Load More button.

feat(various): Now using the useInfiniteLoading hook on the
UpdateAgentView.
Also, updated the associated api request and removed unused code.

fix(ChangeListGroup): Resolved the issue that was related to the
rendering of the StyledEditorName component

fix(various): Stopped using that useModal hook in order to get around
the theme not being passed to the ChangeListGroup component when the
useModal code showed the modal

fix(UpdateAgentView): Quick fix for the setting of the ChangeLog's
previewSize prop

feat(various): The rest of the ChangeListGroup's look is now tied to the theme.

feat(various): This implementation uses the useModal hook to show the ChangeListGroup component within  the associated modal.

fix(various): The ChangeLog modals weren't being given the updated ThemeContext when the theme color was changed.

refactor(useInfiniteLoading): Resolved an eslint issue and removed some console logs

fix(various): Moved the ModalProvider so that is inside of the ThemeProvider; Removed all of the workaround code

fix(UpdateUserView): Fixed a typo

refactor(historicalRecord): Changed the type of the historyType property

fix(various): Made the ChangeLog modals scrollable

feat(various): The ChangeLog modals now open to a static size and their content is now scrollable.
Removed the use of the Modal's scollable props

feat(various): Created a separate class for the changelog's modal body so it doesn't affect the other modals + Now always rendering the footer for the Changelog modal so that scrolling doesn't look wierd

fix(various): Fixed the issue that was causing the main scrollbar to apply to the modals
Solution: Removed the setting of the min-height on the modal-dialog and made use of the Modal's centered flag.

fix(various): The changelog modal's height is now static.

fix(various): The last change item is no longer flickering when more data is being loaded.
Also, added the DimmableContent component to improve the user experience when more data is being loaded.

fix(useInfiniteLoading): Changed useMemo to useEffect because it included a side effect