Modifies <NoContent /> to be more usable. (#657)
NoContent before was hardly more than a style wrapper. This modifies it
to fit it's actual purpose more precisely, which is to display a no
content block for places in the application where there is no content to
display.

Users can use the `extra` prop to display whatever other things they
want in the no content block, such as buttons, forms, etc.