Resolves a merge conflict.
Rebase. This includes the following commits:

Rebase. This includes the following commits:

feat(BitwiseNavbar): Added the base layer of the dropdown

feat(various): Moved the NotificationDropdown into its own file; Added
the header

feat(NotificationDropdown): Set up the rest of the structure for the
NotificationDropdown; Added the state const for the switching between
the unread and read tabs

feat(various): The NotificationDropdown is no longer a part of the NotificationButton; Created a custom tab layout through the use of the React-Bootstrap Tab components

feat(NotificationDropdown): The styling of the tabs are now more in line with the mockup

feat(NotificationDropdown): More styling edits

feat(various): Changed px values to their rem equivalents

feat(BitwiseNavbar and NotificationDropdown): The dropdown's close
button is now functioning.

feat(BitwiseNavbar): Full-screen modal - First Pass Implementation

feat(BitwiseNavbar): Slight fix for the box-shadow

fix(BitwiseNavBar): Sometimes, the notification dropdown would be shown when you clicked on the hamburger icon. This should be the fix for that issue.

feat(BitwiseNavbar): The unread and read tab toggle is now using the correct flex-direction.

feat(NotificationDropdown): The dropdown is now shown above all other components.

fix(BitwiseNavbar): The space to the right and bottom of the notification dropdown is more consistent with their opposite sides when the browser's width or height is decreased.

feat(BitwiseNavbar): The space to the right and bottom of the dropdown are now consistent with their opposite sides when the browser's width or height is decreased.

feat(BitwiseNavbar): Added a max-width to the dropdown on mobile (covers the scenario where someone is playing around with the dimensions of the browser window)

feat(BitwiseNavbar and NotificationDropdown): Made the dropdown look like the other dropdown menus by using the same class
Also, removed some excess white space.

feat(NotificationDropdown): The dropdown's padding is now consistent with the mockup.

feat(NotificationDropdown): Increased the touch target of the close button to a suitable amount; Moved it closer to the edges of the dropdown

feat(various): Now rendering the unread notifications in the notification dropdown

refactor(NotificationDropdown): Changed how the unread notifications are
retrieved; Added the query for the read notifications

feat(NotificationDropdown): The dropdown will now scroll when there isn't enough space to show all of the unread or read notifications.

Additionally, the read notifications are now being rendered.