Resolved a merge conflict
Rebase. This commit includes the following commits:

feat(various): Moved the actions of the agent and user tables into an action dropdown
Also, removed the surrounding divs.

feat(various): Added a CustomToggle; This is used in order to only show the ellipses for the dropdown.

feat(DataTable): The DataTable will now scroll horizontally if there isn't enough space in the viewport to display the entire table.

Revert "feat(DataTable): The DataTable will now scroll horizontally if there isn't enough space in the viewport to display the entire table."
This reverts commit 351b171b7f564365d389e23e26a7b7642c16e869.

feat(various): The look of the dropdown items is now consistent with NavDropdown items (including the hover state).

feat(various): The color of the ellipses is now consistent with the version on the staging server.

feat(DataTable): Added a dropdown button to the right of the columns. The user will be able to use this in order to click on an "Edit Columns" option.

feat(various): Added a hook to get the current width and height of the window (supports resizing); Added a provider in order to pass these values to all of the components in the App tree

feat(various): Added an object that will be the initial state for column visibility for each breakpoint (returned by both of the table data hooks); Also, added relavant types.

feat(various): Now hidding the columns that should be hidden at the xs
breakpoint (as a demo)

feat(various): Now getting the current breakpoint from the useViewport
hook and using that to determine which columns to hide (just need to
update the hidden columns when the breakpoint switches to a new value)

feat(DataTable): Columns will now be hidden accordingly when the
breakpoint's value is changed

feat(DataTable): Now showing a modal with a table that shows the current
values for the visibility for each column based on breakpoint

fix(DataTablie): Resolved the unique key prop issue and made some
parameters in the visibility table more specific

feat(various): Added the xl and xxl breakpoints; Stopped allowing the
column names in the table modal to wrap

feat(various): Now getting the column names dynamically for the table
modal; Also, corrected a type and did some styling edits

feat(DataTable): The visibility states in the table are now toggleable.

refactor(DataTable): Edited how the columns in the table modal were
rendered; Edited the spacing

feat(various): Edited the placement of the CustomToggle button

feat(CustomToggle): Centered the contents of the custom toggle component
and added some padding

refactor(DataTable): Resolved a linter issue and removed some console logs

feat(various): Made some edits to the initial column breakpoint visibility objects

refactor(DataTable): Renamed a method and rearranged some code

feat(DataTable): Corrected some styling issues that were related to the table in the modal and the dropdown menu that shows the Edit Columns dropdown item

refactor(various): Removed the use of the ViewportProvider + Moved the useViewport hook and the related models into the DataTable folder and into relevant subfolders
Also, made relevant changes.

refactor(DataTable): Removed some unnecessary exports