Resolves a merge conflict.
Rebase. This commit includes the following commits:

feat(various): Created a hook out of the NetworkDetector component

feat(various): Refactored the NetworkDetector component so that it shows a barrier if the user goes offline; Moved the use of the useNetworkDetection hook to the NetworkDetector component

feat(notificationApi): Made the notificationApi file consistent with the other api files when it comes to refetching data

feat(store): Data is now being refetched when the user's connection is restored

feat(useReducerInfiniteLoading): In the middle of creating a version of useInfiniteLoading that uses the reducer from useLiveNotifications

feat(various): Added the useNewLiveNotifications hook for testing
purposes and resolves some issues

feat(useReducerInfiniteLoading): The user's unread notifications are now cleared before refetching in the internet reconnection scenario.

fix(useReducerInfiniteLoading): Resolved the issue that was causing the
notifications to be cleared when more notifications are fetched

fix(various): Removed the clearing that was happening when the
notification dropdown was closed. This is already handled by the
useReducerInfiniteLoading hook.

refactor(various): Renamed notification to item in the
useReducerInfiniteLoading hook

refactor(useReducerInfiniteLoading): Implemented a better solution for the duplicate item issue when the user's internet connection is restored

refactor(useReducerInfiniteLoading): Removed some unnecessary code and
added some console logs in order to help identify why this hook keeps
executing

fix(useReducerInfiniteLoading): Resolves the infinite re-rendering of the notification listview

fix(useReducerInfiniteLoading): Corrected the useReducerInfiniteLoading
hook so that it returns the right type for the items and made it be
compatible with the pages that use the useInfiniteLoading hook

refactor(various): Now using the useReducerInfiniteLoading hook in all
of the places where the useInfiniteLoading hook was being used

fix(useReducerInfiniteLoading): Forget to add the error value to the
useMemo dependency list

fix(useReducerInfiniteLoading): The nextItemUrl wasn't being set correctly when the reducer state was reset.

fix(various): I found that the resetApiState call would cause the
infinite loading functionality to just refresh the whole page instead of
working as you would aspect. Removing it fixed this issue. Based on my
testing, I found it wasn't necessary for the notification functionality.

fix(various): It was incorrect to remove the resetApiState function call
for the notification functionality.

refactor(various): Removed the old versions of useLiveNotifications and useInfiniteLoading and replaced them with the new ones

refactored(useLiveNotifications): Removed some commented out code

refactor(NetworkDetector): Removed the InteractionBarrier since the service worker PR will make this functionality unnecessary

refactor(various): Renamed WithNumberIdentifier to WithIdentifier and made its id property accept string values as well

refactor(various): In the middle of making the infinite loading
functionality simpler

refactor(various): Mostly everything is working. Just need to get the
count on the NotificationListView to update when a notification is
removed.

feat(various): The count on the NotificationListView is now being
updated correctly.

refactor(various): Renamed the addOne and addMultiple cases

refactor(useInfinteLoading): Need to implement a better solution for
updating the count on NotificationListView

refactor(various): The count is now being updated correctly again.

refactor(various): Removed unnecessary properties and console logs; Resolved eslint issues