Resolved a merge conflict. Also, added a translation for the message that appears when a user is trying to change their own role.
Rebase. This commit includes the following commits:

Rebase. This commit combines the following commits:

feat(various): Now disabling the role field of the UserDetailForm if the user is updating the user associated with their account

feat(UserDetailForm): Added a message under the Role label when the role field should be disabled
refactor(various): Renamed the shouldRoleBeDisabled prop and used the useMemo hook to make the related code cleaner

feat(various): Made it so that role of admins can't be changed (ex. by other admins)

fix(various): The role of admins can now be changed.